split readonly and index part

+++ ai_librarian/doc_store.py
@@ -1,9 +1,28 @@
 import chromadb
 import numpy as np
+import os
 
 from typing import List, Any
 
 from .base import VectorDocStore, Document, DocId, Embedding
+
+
+class BookStoreFactory:
+    @staticmethod
+    def readonly(book_id, book_dir) -> VectorDocStore:
+        """Get the document store."""
+        collection_name = f"librarian-{book_id}"
+        store_dir = os.path.join(book_dir, "store")
+        return ChromaDocStore.new_local_readonly(
+            collection_name, store_dir
+        )
+
+    @staticmethod
+    def mutable(book_id, book_dir) -> VectorDocStore:
+        """Get the document store."""
+        collection_name = f"librarian-{book_id}"
+        store_dir = os.path.join(book_dir, "store")
+        return ChromaDocStore.new_local(collection_name, store_dir)
 
 
 class ChromaDocStore(VectorDocStore):
@@ -12,7 +31,10 @@ class ChromaDocStore(VectorDocStore):
     client: chromadb.Client
     collection_name: str
 
-    def new_local(collection_name: str, persist_directory: str):
+    @staticmethod
+    def new_local(
+        collection_name: str, persist_directory: str
+    ) -> ChromaDocStore:
         """Create a new ChromaDocStore backed by a local directory."""
         settings = chromadb.config.Settings(
             chroma_db_impl="duckdb+parquet",
@@ -22,13 +44,30 @@ class ChromaDocStore(VectorDocStore):
         client = chromadb.Client(settings)
         return ChromaDocStore(client, collection_name)
 
+    @staticmethod
+    def new_local_readonly(
+        collection_name: str, persistent_directory: str
+    ) -> ChromaDocStore:
+        """Create a readonly ChromaDocStore backed by a local directory."""
+        store = ChromaDocStore.new_local(
+            collection_name, persistent_directory
+        )
+        store.readonly = True
+        return store
+
     def __init__(self, client, collection_name):
         """Create a new ChromaDocStore."""
         self.client = client
         self.collection_name = collection_name
+        self.readonly = False
 
     def collection(self):
         """Get the ChromaDB collection."""
+        if self.readonly:
+            return self.client.get_collection(
+                self.collection_name, embedding_function=dummy_embedding
+            )
+
         return self.client.get_or_create_collection(
             self.collection_name,
             embedding_function=dummy_embedding,
@@ -53,6 +92,9 @@ class ChromaDocStore(VectorDocStore):
 
     def put(self, docs: List[Document]) -> None:
         """Save documents to the store."""
+        if self.readonly:
+            raise Exception("Cannot put documents in a readonly store.")
+
         documents = []
         metadatas = []
         ids = []
@@ -97,6 +139,8 @@ class ChromaDocStore(VectorDocStore):
 
     def reset(self) -> None:
         """Reset the document store."""
+        if self.readonly:
+            raise Exception("Cannot reset a readonly store.")
         self.client.delete_collection(self.collection_name)
 
     def load(self) -> None:
@@ -107,7 +151,15 @@ class ChromaDocStore(VectorDocStore):
 
     def save(self) -> None:
         """Save the document store to disk."""
+        if self.readonly:
+            raise Exception("A readonly store cannot be saved.")
+
         self.client.persist()
+
+    def exists(self) -> bool:
+        """Check if the document store exists."""
+        result = self.collection().peek(n_results=1, include=[])
+        return result["ids"] != []
 
 
 def _results_to_docs(results: Any) -> List[Document]:
@@ -121,10 +173,10 @@ def _results_to_docs(results: Any) -> List[Document]:
 
         embedding = results["embeddings"][i]
         content = results["documents"][i]
-        id = results["ids"][i]
+        id_ = results["ids"][i]
 
         doc = Document(
-            id=id,
+            id=id_,
             content=content,
             metadata=metadata,
             embedding=np.asarray(embedding),
@@ -135,4 +187,4 @@ def _results_to_docs(results: Any) -> List[Document]:
 
 
 def dummy_embedding(_any):
-    raise "Should be unreachable!"
+    raise NotImplementedError("should be unreachable!")
